refactor: resolve a circular dep
