+++ src/renderers/render.ts
@@ -5,27 +5,8 @@ import { type Task } from "@/task/task";
 
 import pkg from "../../package.json";
 import { formatDuration, taskStateIcons } from "../helpers";
-import { type FailureDetails, type TaskState } from "../task/types";
+import { type FailureDetails } from "../task/types";
 import { renderTable } from "./tableRenderer";
-
-/**
- * Colors a status message based on the task state.
- *
- * @param {string} message - The status message to color.
- * @param {TaskState} state - The state of the task ("pending", "running", "success", "failure").
- * @returns {string} - The colored status message.
- */
-export const colorStatusMessage = (message: string, state: TaskState) => {
-  if (!message) {
-    return "";
-  }
-
-  if (state === "failure") {
-    return chalk.red(message);
-  }
-
-  return message;
-};
 
 /**
  * Builds a formatted string summarizing failure details for display.
